requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
pydantic[email]==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9